Snapshot 75.0.3770.156

+++ src/android_webview/renderer/aw_render_frame_ext.h
@@ -59,7 +59,7 @@ class AwRenderFrameExt : public content::RenderFrameObserver {
 
   void OnSetBackgroundColor(SkColor c);
 
-  void OnSmoothScroll(int target_x, int target_y, int duration_ms);
+  void OnSmoothScroll(int target_x, int target_y, uint64_t duration_ms);
 
   void OnSetWillSuppressErrorPage(bool suppress);
