+++ src/content/ppapi_plugin/ppapi_blink_platform_impl.cc
@@ -66,14 +66,13 @@ blink::WebSandboxSupport* PpapiBlinkPlatformImpl::GetSandboxSupport() {
 #endif
 }
 
-unsigned long long PpapiBlinkPlatformImpl::VisitedLinkHash(
-    const char* canonical_url,
-    size_t length) {
+uint64_t PpapiBlinkPlatformImpl::VisitedLinkHash(const char* canonical_url,
+                                                 size_t length) {
   NOTREACHED();
   return 0;
 }
 
-bool PpapiBlinkPlatformImpl::IsLinkVisited(unsigned long long link_hash) {
+bool PpapiBlinkPlatformImpl::IsLinkVisited(uint64_t link_hash) {
   NOTREACHED();
   return false;
 }
